% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{MR_Robin_setup}
\alias{MR_Robin_setup}
\title{Set up data for MR-Robin algorithm.}
\usage{
MR_Robin_setup(eqtl_data, gwas_data, LD, snpID, geneID, ncond)
}
\arguments{
\item{eqtl_data}{data.frame of summary statistics from eQTL study.}

\item{gwas_data}{data.frame of summary statistics from GWAS study.}

\item{LD}{matrix of LD correlation coefficients.}

\item{snpID}{vector of variant identifiers be used as instruments for \code{geneID}.}

\item{geneID}{character string of the gene to be tested.}

\item{ncond}{integer of the number of conditions (e.g. tissues) analyzed by the eQTL study.}
}
\value{
A list with the following elements:

\tabular{ll}{
\code{eqtl_betas} \tab matrix of coefficient estimates (betas) from eQTL study.\cr
\code{eqtl_se} \tab matrix of standard errors for coefficient estimates from eQTL study.\cr
\code{gwas_betas} \tab vector of coefficient estimates (betas) from GWAS study.\cr
\code{gwas_se} \tab vector of standard errors for coefficient estimates from GWAS study.\cr
\code{LD} \tab matrix of LD correlation coefficients.\cr
\code{snpID} \tab vector of variant identifiers to be used as instrumental variables.\cr
}

The returned list elements can be passed to the main \code{\link{MR_Robin}} function.
}
\description{
Sets up data to run the main MR-Robin algorithm:
a two-sample Mendelian Randomization method
ROBust to correlated and some INvalid instruments.
}
\details{
The following are additional details describing the input arguments.

The data.frame \code{eqtl_data} should have the character
variables \code{gene_id} and \code{variant_id} (as identifiers), as well as the variables
\code{beta_j} and \code{SE_j} for \eqn{j} in \{1,...,\code{ncond}\}, corresponding to the coefficient and
standard error estimates from each respective eQTL analysis.

The data.frame \code{gwas_data} should have the character
variable \code{variant_id}, as well as the variables
\code{beta} and \code{SE} corresponding to the coefficient and
standard error estimates from the GWAS analysis.

Note that the matrix \code{LD} should hold \emph{correlation coefficients}
(i.e. \eqn{r}), not their squared values (\eqn{r^2}).
}
