% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MrRobin.R
\docType{package}
\name{MrRobin}
\alias{MrRobin}
\alias{MrRobin-package}
\title{MrRobin: A package to conduct two-sample Mendelian Randomization analysis.}
\description{
The MrRobin package implements the two-sample Mendelian Randomization
method ROBust to correlated and some INvalid instruments, a.k.a. 'MR-Robin'.
}
\section{Main}{

To run the random slope reverse regression mixed model of MR-Robin,
call the function \code{\link{MR_Robin}}.
}

