% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MR_Robin}
\alias{MR_Robin}
\title{Conduct two-sample Mendelian Randomization using summary statistics}
\usage{
MR_Robin(eqtl_betas, eqtl_se, gwas_betas, gwas_se, LD, snpID)
}
\arguments{
\item{eqtl_betas}{matrix of coefficient estimates (betas) from eQTL study.}

\item{eqtl_se}{matrix of standard errors for coefficient estimates from eQTL study.}

\item{gwas_betas}{vector of coefficient estimates (betas) from GWAS study.}

\item{gwas_se}{vector of standard errors for coefficient estimates from GWAS study.}

\item{LD}{matrix of LD correlation coefficients (\eqn{r}, not \eqn{r^2}).}

\item{snpID}{vector of variant identifiers to be used as instrumental variables.}
}
\value{
An object of class \code{lmerMod}, returned from the reverse regression random slope
mixed model run by MR-Robin.

To conduct inference on the returned results, use function \code{\link{MR_Robin_resample}}.
}
\description{
Runs the main MR-Robin algorithm: a two-sample Mendelian Randomization method
ROBust to correlated and some INvalid instruments.
}
\details{
The following are additional details describing the input arguments.
For \code{eqtl_betas} and \code{eqtl_se}, each row \eqn{i} corresponds to a SNP/variant
while each column \eqn{j} holds the summary statistics in condition \eqn{j} (e.g. a tissue type).
Note that the matrix \code{LD} should hold \emph{correlation coefficients}
(i.e. \eqn{r}), not their squared values (\eqn{r^2}).
}
